+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This project does *not* adhere to [Semantic Versioning](https://semver.org).
 ## [Unreleased](https://github.com/dsa-ou/allowed/compare/v1.3.0...HEAD)
 These changes are in the GitHub repository but not on [PyPI](https://pypi.org/project/allowed).
 
+### Added
+- option `-v`/ `--verbose`: display additional info about the checking process
+
 ### Fixed
 - report syntax errors in notebooks in the same way as in Python files